+++ test/src/level_converter/log_level_to_ansi_pen_converter_test.dart
@@ -79,8 +79,8 @@ void main() {
         converter.shout?.down,
         (AnsiPen()..magenta(bg: true, bold: true)).down,
       );
-      expect(converter.defaultValue.up, '');
-      expect(converter.defaultValue.down, '');
+      expect(converter.defaultValue.up, AnsiPen().up);
+      expect(converter.defaultValue.down, AnsiPen().down);
     });
   });
 }