+++ CHANGELOG.md
@@ -2,6 +2,8 @@
 
 - feat: added `IoLogSink`
 - deprecated `PrintSink`
+- fix: json log record parser would add stackTrace even when none is present
+- fix: a bunch of tests
 
 ## 0.5.3